添加ViewPager2
1. 使用 ViewPager2 代替 RecyclerView。
2. 将 VideoPlayerActivity 中的内容移至 MyViewHolder 内。
